@@ -84,7 +84,8 @@ def prepare_for_mms(
     image_uri: str,
     inference_spec: InferenceSpec = None,
 ) -> str:
-    """Prepares for InferenceSpec using model_path, writes inference.py, and captures dependencies to generate secret_key.
+    """Prepares for InferenceSpec using model_path, writes inference.py, \
+        and captures dependencies to generate secret_key.
 
     Args:to
         model_path (str) : Argument
