@@ -720,8 +720,8 @@ def _generate_processor(
     )
     return FrameworkProcessor(
         estimator_cls=PyTorch,
-        framework_version="1.10",
-        py_version="py38",
+        framework_version="1.13.1",
+        py_version="py39",
         instance_count=1,
         instance_type="ml.m5.xlarge",
         role=execution_role,
