@@ -1388,7 +1388,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
         mock_sagemaker_session.list_tags = mock_list_tags
         mock_list_tags.return_value = [{"Key": "blah", "Value": "blah1"}]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, None, None),
         )
@@ -1403,7 +1403,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.MODEL_ID, "Value": "model_id"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             ("model_id", None, None, None),
         )
@@ -1418,7 +1418,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.MODEL_VERSION, "Value": "model_version"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, "model_version", None, None),
         )
@@ -1430,7 +1430,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
         mock_sagemaker_session.list_tags = mock_list_tags
         mock_list_tags.return_value = [{"Key": "blah", "Value": "blah1"}]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, None, None),
         )
@@ -1445,7 +1445,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.INFERENCE_CONFIG_NAME, "Value": "config_name"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, "config_name", None),
         )
@@ -1460,7 +1460,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.TRAINING_CONFIG_NAME, "Value": "config_name"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, None, "config_name"),
         )
@@ -1476,7 +1476,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.TRAINING_CONFIG_NAME, "Value": "training_config_name"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, "inference_config_name", "training_config_name"),
         )
@@ -1492,7 +1492,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.MODEL_VERSION, "Value": "model_version"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             ("model_id", "model_version", None, None),
         )
@@ -1510,7 +1510,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.MODEL_VERSION, "Value": "model_version_2"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, None, None),
         )
@@ -1528,7 +1528,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": random.choice(EXTRA_MODEL_VERSION_TAGS), "Value": "model_version_1"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             ("model_id_1", "model_version_1", None, None),
         )
@@ -1546,7 +1546,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": random.choice(EXTRA_MODEL_VERSION_TAGS), "Value": "model_version_2"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             (None, None, None, None),
         )
@@ -1564,7 +1564,7 @@ class TestGetModelIdVersionFromResourceArn(TestCase):
             {"Key": JumpStartTag.INFERENCE_CONFIG_NAME, "Value": "config_name_2"},
         ]
 
-        self.assertEquals(
+        self.assertEqual(
             utils.get_jumpstart_model_info_from_resource_arn("some-arn", mock_sagemaker_session),
             ("model_id_1", "model_version_1", None, None),
         )
