@@ -1049,15 +1049,16 @@ def test_model_trainer_local_full_init(
 
     model_trainer.train()
 
-    assert mock_local_container.train.called_once_with(
+    mock_local_container.assert_called_once_with(
         training_job_name=unique_name,
         instance_type=compute.instance_type,
         instance_count=compute.instance_count,
         image=training_image,
         container_root=local_container_root,
         sagemaker_session=modules_session,
-        container_entry_point=DEFAULT_ENTRYPOINT,
+        container_entrypoint=DEFAULT_ENTRYPOINT,
         container_arguments=DEFAULT_ARGUMENTS,
+        input_data_config=ANY,
         hyper_parameters=hyperparameters,
         environment=environment,
     )
