@@ -21,8 +21,8 @@ from sagemaker.serve.detector.dependency_manager import _parse_dependency_list, 
 
 DEPENDENCY_LIST = [
     "requests==2.26.0",
-    "numpy>=2.0.0,<2.3.0",
-    "pandas==2.2.3",
+    "numpy==1.26.4",
+    "pandas<=1.3.3",
     "matplotlib<3.5.0",
     "scikit-learn>0.24.1",
     "Django!=4.0.0",
@@ -34,8 +34,8 @@ DEPENDENCY_LIST = [
 
 EXPECTED_DEPENDENCY_MAP = {
     "requests": "==2.26.0",
-    "numpy": "==2.0.0",
-    "pandas": "==2.2.3",
+    "numpy": "==1.26.4",
+    "pandas": "<=1.3.3",
     "matplotlib": "<3.5.0",
     "scikit-learn": ">0.24.1",
     "Django": "!=4.0.0",
