@@ -91,8 +91,8 @@ class MultiModelServerPrepareTests(TestCase):
         mock_disk_space.assert_called_once_with(mock_model_path)
         mock_disk_usage.assert_called_once()
 
-        self.assertEquals(ret_model_path, mock_model_path)
-        self.assertEquals(ret_code_dir, mock_code_dir)
+        self.assertEqual(ret_model_path, mock_model_path)
+        self.assertEqual(ret_code_dir, mock_code_dir)
 
     @patch("sagemaker.serve.model_server.multi_model_server.prepare.Path")
     def test_create_dir_structure_invalid_path(self, mock_path):
@@ -104,4 +104,4 @@ class MultiModelServerPrepareTests(TestCase):
         with self.assertRaises(ValueError) as context:
             _create_dir_structure(mock_model_path)
 
-        self.assertEquals("model_dir is not a valid directory", str(context.exception))
+        self.assertEqual("model_dir is not a valid directory", str(context.exception))
