@@ -50,8 +50,8 @@ class TgiPrepareTests(TestCase):
         mock_disk_space.assert_called_once_with(mock_model_path)
         mock_disk_usage.assert_called_once()
 
-        self.assertEquals(ret_model_path, mock_model_path)
-        self.assertEquals(ret_code_dir, mock_code_dir)
+        self.assertEqual(ret_model_path, mock_model_path)
+        self.assertEqual(ret_code_dir, mock_code_dir)
 
     @patch("sagemaker.serve.model_server.tgi.prepare.Path")
     def test_create_dir_structure_invalid_path(self, mock_path):
@@ -63,7 +63,7 @@ class TgiPrepareTests(TestCase):
         with self.assertRaises(ValueError) as context:
             _create_dir_structure(mock_model_path)
 
-        self.assertEquals("model_dir is not a valid directory", str(context.exception))
+        self.assertEqual("model_dir is not a valid directory", str(context.exception))
 
     @patch("sagemaker.serve.model_server.tgi.prepare.S3Downloader")
     @patch("builtins.open", read_data="data")
