@@ -99,7 +99,7 @@ def test_pipeline_create_and_update_with_config_injection(sagemaker_session_mock
         RoleArn=pipeline_role_arn,
     )
     pipeline.upsert()
-    assert sagemaker_session_mock.sagemaker_client.update_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.update_pipeline.assert_called_with(
         PipelineName="MyPipeline",
         PipelineDefinition=pipeline.definition(),
         RoleArn=pipeline_role_arn,
@@ -130,7 +130,7 @@ def test_pipeline_create_with_parallelism_config(sagemaker_session_mock, role_ar
         role_arn=role_arn,
         parallelism_config=dict(MaxParallelExecutionSteps=10),
     )
-    assert sagemaker_session_mock.sagemaker_client.create_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.create_pipeline.assert_called_with(
         PipelineName="MyPipeline",
         PipelineDefinition=pipeline.definition(),
         RoleArn=role_arn,
@@ -149,7 +149,7 @@ def test_pipeline_create_and_start_with_parallelism_config(sagemaker_session_moc
         role_arn=role_arn,
         parallelism_config=dict(MaxParallelExecutionSteps=10),
     )
-    assert sagemaker_session_mock.sagemaker_client.create_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.create_pipeline.assert_called_with(
         PipelineName="MyPipeline",
         PipelineDefinition=pipeline.definition(),
         RoleArn=role_arn,
@@ -168,7 +168,7 @@ def test_pipeline_create_and_start_with_parallelism_config(sagemaker_session_moc
 
     # Specify ParallelismConfiguration to another value which will be honored in backend
     pipeline.start(parallelism_config=dict(MaxParallelExecutionSteps=20))
-    assert sagemaker_session_mock.sagemaker_client.start_pipeline_execution.called_with(
+    sagemaker_session_mock.sagemaker_client.start_pipeline_execution.assert_called_with(
         PipelineName="MyPipeline",
         ParallelismConfiguration={"MaxParallelExecutionSteps": 20},
     )
@@ -209,7 +209,7 @@ def test_pipeline_update(sagemaker_session_mock, role_arn):
     assert not pipeline.steps
     pipeline.update(role_arn=role_arn)
     assert len(json.loads(pipeline.definition())["Steps"]) == 0
-    assert sagemaker_session_mock.sagemaker_client.update_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.update_pipeline.assert_called_with(
         PipelineName="MyPipeline", PipelineDefinition=pipeline.definition(), RoleArn=role_arn
     )
 
@@ -253,7 +253,7 @@ def test_pipeline_update(sagemaker_session_mock, role_arn):
 
     pipeline.update(role_arn=role_arn)
     assert len(json.loads(pipeline.definition())["Steps"]) == 3
-    assert sagemaker_session_mock.sagemaker_client.update_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.update_pipeline.assert_called_with(
         PipelineName="MyPipeline", PipelineDefinition=pipeline.definition(), RoleArn=role_arn
     )
 
@@ -345,7 +345,11 @@ def test_pipeline_update_with_parallelism_config(sagemaker_session_mock, role_ar
         role_arn=role_arn,
         parallelism_config=dict(MaxParallelExecutionSteps=10),
     )
-    assert sagemaker_session_mock.sagemaker_client.update_pipeline.called_with(
+    pipeline.update(
+        role_arn=role_arn,
+        parallelism_config={"MaxParallelExecutionSteps": 10},
+    )
+    sagemaker_session_mock.sagemaker_client.update_pipeline.assert_called_with(
         PipelineName="MyPipeline",
         PipelineDefinition=pipeline.definition(),
         RoleArn=role_arn,
@@ -418,13 +422,11 @@ def test_pipeline_upsert_resource_already_exists(sagemaker_session_mock, role_ar
     sagemaker_session_mock.sagemaker_client.update_pipeline.assert_called_once_with(
         PipelineName="MyPipeline", PipelineDefinition=pipeline.definition(), RoleArn=role_arn
     )
-    assert sagemaker_session_mock.sagemaker_client.list_tags.called_with(
-        ResourceArn="mock_pipeline_arn"
-    )
+    sagemaker_session_mock.sagemaker_client.list_tags.assert_called_with(ResourceArn="pipeline-arn")
 
     tags.append({"Key": "dummy", "Value": "dummy_tag"})
-    assert sagemaker_session_mock.sagemaker_client.add_tags.called_with(
-        ResourceArn="mock_pipeline_arn", Tags=tags
+    sagemaker_session_mock.sagemaker_client.add_tags.assert_called_with(
+        ResourceArn="pipeline-arn", Tags=tags
     )
 
 
@@ -523,7 +525,7 @@ def test_pipeline_delete(sagemaker_session_mock):
         sagemaker_session=sagemaker_session_mock,
     )
     pipeline.delete()
-    assert sagemaker_session_mock.sagemaker_client.delete_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.delete_pipeline.assert_called_with(
         PipelineName="MyPipeline",
     )
 
@@ -536,7 +538,7 @@ def test_pipeline_describe(sagemaker_session_mock):
         sagemaker_session=sagemaker_session_mock,
     )
     pipeline.describe()
-    assert sagemaker_session_mock.sagemaker_client.describe_pipeline.called_with(
+    sagemaker_session_mock.sagemaker_client.describe_pipeline.assert_called_with(
         PipelineName="MyPipeline",
     )
 
@@ -552,17 +554,17 @@ def test_pipeline_start(sagemaker_session_mock):
         sagemaker_session=sagemaker_session_mock,
     )
     pipeline.start()
-    assert sagemaker_session_mock.start_pipeline_execution.called_with(
+    sagemaker_session_mock.sagemaker_client.start_pipeline_execution.assert_called_with(
         PipelineName="MyPipeline",
     )
 
     pipeline.start(execution_display_name="pipeline-execution")
-    assert sagemaker_session_mock.start_pipeline_execution.called_with(
+    sagemaker_session_mock.sagemaker_client.start_pipeline_execution.assert_called_with(
         PipelineName="MyPipeline", PipelineExecutionDisplayName="pipeline-execution"
     )
 
     pipeline.start(parameters=dict(alpha="epsilon"))
-    assert sagemaker_session_mock.start_pipeline_execution.called_with(
+    sagemaker_session_mock.sagemaker_client.start_pipeline_execution.assert_called_with(
         PipelineName="MyPipeline", PipelineParameters=[{"Name": "alpha", "Value": "epsilon"}]
     )
 
@@ -821,10 +823,8 @@ def test_pipeline_build_parameters_from_execution(sagemaker_session_mock):
         pipeline_execution_arn=reference_execution_arn,
         parameter_value_overrides=parameter_value_overrides,
     )
-    assert (
-        sagemaker_session_mock.sagemaker_client.list_pipeline_parameters_for_execution.called_with(
-            PipelineExecutionArn=reference_execution_arn
-        )
+    sagemaker_session_mock.sagemaker_client.list_pipeline_parameters_for_execution.assert_called_with(
+        PipelineExecutionArn=reference_execution_arn
     )
     assert len(parameters) == 1
     assert parameters["TestParameterName"] == "NewParameterValue"
@@ -850,10 +850,8 @@ def test_pipeline_build_parameters_from_execution_with_invalid_overrides(sagemak
         + f"are not present in the pipeline execution: {reference_execution_arn}"
         in str(error)
     )
-    assert (
-        sagemaker_session_mock.sagemaker_client.list_pipeline_parameters_for_execution.called_with(
-            PipelineExecutionArn=reference_execution_arn
-        )
+    sagemaker_session_mock.sagemaker_client.list_pipeline_parameters_for_execution.assert_called_with(
+        PipelineExecutionArn=reference_execution_arn
     )
 
 
@@ -908,24 +906,23 @@ def test_pipeline_execution_basics(sagemaker_session_mock):
     )
     execution = pipeline.start()
     execution.stop()
-    assert sagemaker_session_mock.sagemaker_client.stop_pipeline_execution.called_with(
+    sagemaker_session_mock.sagemaker_client.stop_pipeline_execution.assert_called_with(
         PipelineExecutionArn="my:arn"
     )
     execution.describe()
-    assert sagemaker_session_mock.sagemaker_client.describe_pipeline_execution.called_with(
+    sagemaker_session_mock.sagemaker_client.describe_pipeline_execution.assert_called_with(
         PipelineExecutionArn="my:arn"
     )
     steps = execution.list_steps()
-    assert sagemaker_session_mock.sagemaker_client.describe_pipeline_execution_steps.called_with(
+    sagemaker_session_mock.sagemaker_client.list_pipeline_execution_steps.assert_called_with(
         PipelineExecutionArn="my:arn"
     )
     assert len(steps) == 1
     list_parameters_response = execution.list_parameters()
-    assert (
-        sagemaker_session_mock.sagemaker_client.list_pipeline_parameters_for_execution.called_with(
-            PipelineExecutionArn="my:arn"
-        )
+    sagemaker_session_mock.sagemaker_client.list_pipeline_parameters_for_execution.assert_called_with(
+        PipelineExecutionArn="my:arn"
     )
+
     parameter_list = list_parameters_response["PipelineParameters"]
     assert len(parameter_list) == 1
     assert parameter_list[0]["Name"] == "TestParameterName"
